Allow multiple outputs to use same animated image
Multiple outputs can use the same image, but a single display is in
charge of driving the animation. When the animation results in a new
frame, the image increments a frame counter. All outputs can compare
this count with their own count to determine if a new frame is to be
rendered.